Updated Client Window
Most everything is in place except for connecting and sending to server.
Need to add place for user to enter ip + port. Can currently open and
edit files and save them. Client in place to begin sending files within
a day or so.

Code is a little bit messy, will be cleaned up later!