Client connection
Client pretty much good to go. Can open, save, save as, edit a file, and
submit/connect to a server. Currently tested using HW45 server, works
like a charm. Does not (as of yet) send the file over, but that is easy
enough and only needs to send it to the server.